% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist.R
\name{as.data.frame.sample_dist}
\alias{as.data.frame.sample_dist}
\title{Convert sampling distribution to a data frame}
\usage{
\method{as.data.frame}{sample_dist}(x, ..., types = NULL)
}
\arguments{
\item{x}{sample_dist object to convert}

\item{...}{further parameters supplied to as.data.frame}

\item{types}{the sequence types to extract. Defaults to NULL (all types)}
}
\value{
a data.frame with columns type, iteration, source and log_p
}
\description{
Convert sampling distribution to a data frame
}
