% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{attribution_ar1}
\alias{attribution_ar1}
\title{Attribute cases to sources via MCMC with an AR(1) model on time}
\usage{
attribution_ar1(formula, time, sampling_dist, data, iterations = 10000,
  burnin = 1000, thinning = 100, priors = NULL)
}
\arguments{
\item{formula}{A formula of the form `GenoType ~ Covariates` for cases}

\item{time}{The name of the time column}

\item{sampling_dist}{A sampling_dist object previously fitted to source genotypes}

\item{data}{An optional data frame to take the variables in `formula` from}

\item{iterations}{the number of iterations to use in the MCMC}

\item{burnin}{the number of iterations to eliminate due to burnin}

\item{thinning}{how many iterations to perform before taking a sample}

\item{priors}{A list specifying the priors to use.}
}
\value{
an object of type attribution
}
\description{
Attribute cases to sources via MCMC with an AR(1) model on time
}
\seealso{
print.attribution, summary.attribution, plot.attribution
}
