% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution_intermediate.R
\name{attribution_with_intermediate}
\alias{attribution_with_intermediate}
\title{Attribute cases to sources via an intermediate source with MCMC}
\usage{
attribution_with_intermediate(
  formula,
  int_genotypes,
  sampling_dist,
  data,
  iterations = 10000,
  burnin = 1000,
  thinning = 100,
  priors = NULL
)
}
\arguments{
\item{formula}{A formula of the form `GenoType ~ Covariates` for human 'sink' cases}

\item{int_genotypes}{A vector of intermediate genotypes (e.g. from water or similar)}

\item{sampling_dist}{A sampling_dist object previously fitted to source genotypes (not intermediate)}

\item{data}{An optional data frame to take the variables in `formula` from}

\item{iterations}{the number of iterations to use in the MCMC}

\item{burnin}{the number of iterations to eliminate due to burnin}

\item{thinning}{how many iterations to perform before taking a sample}

\item{priors}{a list of priors to use of the form `priors=list(theta=list(mean=0,prec=0.1))`}
}
\value{
an object of type intermediate_attribution
}
\description{
Attribute cases to sources via an intermediate source with MCMC
}
\seealso{
print.attribution, summary.attribution, plot.attribution
}
