% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{DIC.attribution}
\alias{DIC.attribution}
\title{Retrieve the Deviance Information Critieria (DIC) of an attribution object}
\usage{
\method{DIC}{attribution}(x)
}
\arguments{
\item{x}{An object of class `attribution`}
}
\value{
The DIC of the model
}
\description{
Retrieve the Deviance Information Critieria (DIC) of an attribution object
}
