% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{DIC}
\alias{DIC}
\title{Generic for retrieving the DIC from a model fit}
\usage{
DIC(x)
}
\arguments{
\item{x}{an object}
}
\value{
the DIC from the model fit
}
\description{
Generic for retrieving the DIC from a model fit
}
