% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genotype_attribution}
\alias{genotype_attribution}
\title{Genotype attribution dataset}
\format{A data frame containing attribution information
\describe{
  \item{ST}{The sequence type identifier}
  \item{ASP}{The ASP housekeeping gene}
  \item{GLN}{The GLN housekeeping gene}
  \item{GLT}{The GLT housekeeping gene}
  \item{GLY}{The GLY housekeeping gene}
  \item{PGM}{The PGM housekeeping gene}
  \item{TKT}{The TKT housekeeping gene}
  \item{UNC}{The UNC housekeeping gene}
  \item{Poultry}{The probability the sample is from poultry}
  \item{Ruminants}{The probability the sample is from ruminants}
  \item{Water}{The probability the sample is from environmental water}
  \item{Other}{The probability the sample is from other sources}
  \item{Iteration}{The posterior sample iteration}
}}
\usage{
genotype_attribution
}
\description{
A dataset containing posterior samples of attribution to sources for a list
of genotypes.
}
\keyword{datasets}

