% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.R
\name{get_num_samples}
\alias{get_num_samples}
\title{Get the number of posterior samples available for attribution}
\usage{
get_num_samples()
}
\description{
Get the number of posterior samples available for attribution
}

