% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic.R
\name{get_source_probability_sample}
\alias{get_source_probability_sample}
\title{Retrieve the i-th posterior sample for attribution for a given genotype}
\usage{
get_source_probability_sample(genotype, sample = NULL)
}
\arguments{
\item{genotype}{the genotype to retrieve.}

\item{sample}{the sample to retrieve. Defaults to NULL, where all samples will be retrieved.}
}
\value{
A data frame of sequence types, their allelic profile, and clonal complex.
}
\description{
Retrieve the i-th posterior sample for attribution for a given genotype
}
\seealso{
pubmlst
}

