% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{get_variable}
\alias{get_variable}
\title{Extract a variable from the posterior of an attribution fit}
\usage{
get_variable(x, variable)
}
\arguments{
\item{x}{an attribution object, fit using attribution.}

\item{variable}{a character string specifying the variable to extract.}
}
\value{
a list of posterior values for the given variable
}
\description{
Extract a variable from the posterior of an attribution fit
}

