% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{iterations.attribution}
\alias{iterations.attribution}
\title{Retrieve the number of MCMC samples from an attribution object}
\usage{
\method{iterations}{attribution}(x)
}
\arguments{
\item{x}{An object of class `attribution`}
}
\value{
The number of iterations (integer)
}
\description{
Retrieve the number of MCMC samples from an attribution object
}
