% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist.R
\name{iterations}
\alias{iterations}
\title{Generic for retrieving the number of iterations from an object}
\usage{
iterations(x)
}
\arguments{
\item{x}{an object}
}
\value{
the number of iterations
}
\description{
Generic for retrieving the number of iterations from an object
}

