% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist.R
\name{iterations.sample_dist}
\alias{iterations.sample_dist}
\title{Retrieve the number of iterations from a sample_dist object}
\usage{
\method{iterations}{sample_dist}(x)
}
\arguments{
\item{x}{sample_dist object}
}
\value{
the number of iterations
}
\description{
Retrieve the number of iterations from a sample_dist object
}

