% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{MCMC parameters
time column (in case time is important)
design matrix
parameter vector
precision, auto-correlation
initialise p
compute log-likelihood for each covariate pattern}
\usage{
mcmc(humans, x, formula, phi, iterations = 10000, burnin = 1000)
}
\description{
MCMC parameters
time column (in case time is important)
design matrix
parameter vector
precision, auto-correlation
initialise p
compute log-likelihood for each covariate pattern
}

