% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_no_ar1}
\alias{mcmc_no_ar1}
\title{parameter vector
initialise p
compute log-likelihood for each covariate pattern}
\usage{
mcmc_no_ar1(humans, X, phi, iterations = 10000, burnin = 1000,
  thinning = 100)
}
\description{
parameter vector
initialise p
compute log-likelihood for each covariate pattern
}

