% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{plot.attribution}
\alias{plot.attribution}
\title{Plot an attribution object}
\usage{
\method{plot}{attribution}(x, ...)
}
\arguments{
\item{x}{An object of class `attribution`}

\item{...}{further parameters passed to lower-level plot commands.}
}
\description{
Plot an attribution object
}

