% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist.R
\name{plot.sample_dist}
\alias{plot.sample_dist}
\title{Produce a summary plot (stacked barplot) of posterior means}
\usage{
\method{plot}{sample_dist}(x, ...)
}
\arguments{
\item{x}{sample_dist object to plot}

\item{...}{further parameters supplied to plot}
}
\description{
Produce a summary plot (stacked barplot) of posterior means
}
