% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{predict.attribution}
\alias{predict.attribution}
\title{Predict attribution on a new data set}
\usage{
\method{predict}{attribution}(object, newdata = NULL, FUN = median, ...)
}
\arguments{
\item{object}{an object of class `attribution`, usually a result of a call to `attribution`.}

\item{newdata}{a `data.frame` to predict attribution values. Set to NULL to use the reduced model matrix from the fitted model.}

\item{FUN}{a function to operate on the posterior attribution, defaults to `median`, use `identity` to retrieve all posterior samples.}

\item{...}{further parameters to pass to FUN}
}
\value{
a data frame containing the posterior attribution (p), the source (Source), the covariate
        combination (X), and if multiple results are returned from FUN, an additional column named
        after the function call.
}
\description{
Predict attribution on a new data set
}

