% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{print.attribution}
\alias{print.attribution}
\title{Print an overview of an attribution object}
\usage{
\method{print}{attribution}(x)
}
\arguments{
\item{x}{An object of class `attribution`}
}
\description{
Print an overview of an attribution object
}

