% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist.R
\name{print.sample_dist}
\alias{print.sample_dist}
\title{Print a sample_dist object}
\usage{
\method{print}{sample_dist}(x, ...)
}
\arguments{
\item{x}{sample_dist object to print}

\item{...}{further parameters supplied to the print function}
}
\description{
Print a sample_dist object
}

