% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist.R
\name{st_fit}
\alias{st_fit}
\title{Fit the sampling distribution of genotypes to sources using different models.}
\usage{
st_fit(formula, non_primary = "Human", method = "island", data, ...)
}
\arguments{
\item{formula}{A formula of the form Source ~ Genotype}

\item{non_primary}{one or more sources that should be considered 'output' sources. No genotype distribution is computed for these,
but P(ST | source) will be computed for these STs in addition to those observed on other sources.}

\item{method}{the method to use to fit the genotype distribution. Only \code{"island"} is supported currently.}

\item{data}{optional data frame from which to take variables in \code{formula}.}

\item{...}{further paramters to pass to the method-specific fitting algorithms.}
}
\value{
an object derived from class sampling_dist.
}
\description{
Fit the sampling distribution of genotypes to sources using different models.
}
\seealso{
\code{\link{st_fit_island}}, \code{\link{print.sample_dist}}, \code{\link{plot.sample_dist}}, \code{\link{summary.sample_dist}}
}

