% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist_dirichlet.R
\name{st_fit_dirichlet}
\alias{st_fit_dirichlet}
\title{Fit the sampling distribution of genotypes to sources using a Dirichlet/multinomial distribution.}
\usage{
st_fit_dirichlet(formula, prior = 1, non_primary = "Human", samples = 100,
  data)
}
\arguments{
\item{formula}{A formula of the form Source ~ Genotype}

\item{prior}{The Dirichlet prior. Currently only a constant. Defaults to 1.}

\item{non_primary}{one or more sources that should be considered 'output' sources. No genotype distribution is computed for these,
but P(ST | source) will be computed for these STs in addition to those observed on other sources.}

\item{samples}{the number of iterations to sample. Defaults to 100.}

\item{data}{optional data frame from which to take variables in \code{formula} and \code{sequence}.}
}
\value{
an object of class dirichlet which derives from sampling_dist.
}
\description{
This routine estimates the sampling distribution of the genotypes on each
source by assuming the observed counts arise from a multinomial distribution,
with a Dirichlet prior on the parameter vector, giving a Dirichlet posterior.
Once estimated, the sampling distribution may be used with
\code{\link{attribution}} to attribute cases to sources based on genotype data in addition to
other covariates on the cases.
}
\seealso{
\code{\link{st_fit}}, \code{\link{print.sample_dist}}, \code{\link{plot.sample_dist}}, \code{\link{summary.sample_dist}}
}
