% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist_island.R
\name{st_fit_island}
\alias{st_fit_island}
\title{Fit the sampling distribution of genotypes to sources using the assymmetric island model.}
\usage{
st_fit_island(formula, sequences, non_primary = "Human", iters = 10000,
  data)
}
\arguments{
\item{formula}{A formula of the form Source ~ Genotype}

\item{sequences}{A formula of the form ~ASP + GLN + GLT specifying columns for the allelic profile of each type}

\item{non_primary}{one or more sources that should be considered 'output' sources. No genotype distribution is computed for these,
but P(ST | source) will be computed for these STs in addition to those observed on other sources.}

\item{iters}{the number of iterations to sample. An additional burnin period of 10% of this value will be used, and
100 samples will be taken (or every sample, whichever is smaller).}

\item{data}{optional data frame from which to take variables in \code{formula} and \code{sequence}.}
}
\value{
an object of class island which derives from sampling_dist.
}
\description{
This routine uses the asymmetric island model (See \href{http://dx.doi.org/10.1371/journal.pgen.1000203}{D. Wilson et al 2008}) to estimate the sampling distribution
of genotypes on each source. Once estimated, the sampling distribution may be used with
\code{\link{attribution}} to attribute cases to sources based on genotype data in addition to
other covariates on the cases.
}
\seealso{
\code{\link{st_fit}}, \code{\link{print.sample_dist}}, \code{\link{plot.sample_dist}}, \code{\link{summary.sample_dist}}
}
