% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{summary.attribution}
\alias{summary.attribution}
\title{Retrieve a summary of an attribution object}
\usage{
\method{summary}{attribution}(x)
}
\arguments{
\item{x}{an object of class `attribution`, usually, a result of a call to `attribution`}
}
\description{
Retrieve a summary of an attribution object
}

