% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist.R
\name{summary.sample_dist}
\alias{summary.sample_dist}
\title{Print a summary for a sample_dist object}
\usage{
\method{summary}{sample_dist}(x, ...)
}
\arguments{
\item{x}{sample_dist object to summarise}

\item{...}{further parameters supplied to summary}
}
\value{
posterior means of the ST distribution on each source
}
\description{
Print a summary for a sample_dist object
}
