% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{manawatu}
\alias{manawatu}
\title{Example human *Campylobacter jejuni* dataset from the Manawatu, New Zealand}
\format{A data frame containing attribution information
\describe{
  \item{ST}{The sequence type identifier}
  \item{ASP}{The ASP housekeeping gene}
  \item{GLN}{The GLN housekeeping gene}
  \item{GLT}{The GLT housekeeping gene}
  \item{GLY}{The GLY housekeeping gene}
  \item{PGM}{The PGM housekeeping gene}
  \item{TKT}{The TKT housekeeping gene}
  \item{UNC}{The UNC housekeeping gene}
  \item{Source}{The source (Human, Poultry, Ruminants, Water, Other)}
  \item{Location}{A factor containing Urban or Rural for each Human case}
  \item{Intervention}{A factor containing before or after for each Human case, specifying whether
  it occurred before or after the 2007-2008 intervention in the Poultry industry}
}}
\usage{
manawatu
}
\description{
A dataset containing MLST isolates from human and animal sources from the Manawatu,
New Zealand.
}
\keyword{datasets}

