% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{print.attribution}
\alias{print.attribution}
\title{Print an overview of an attribution object}
\usage{
\method{print}{attribution}(x, ...)
}
\arguments{
\item{x}{an object of class `attribution`}

\item{...}{further parameters passed to base print functions}
}
\description{
Print an overview of an attribution object
}

